Use int instead of ACCESS_STATUS as key type in class CacheController

+++ CacheController.cpp
@@ -6,14 +6,14 @@ CacheController::CacheController(const CacheConfig& L1_config, const CacheConfig
                                  unsigned int L1_latency, unsigned int L2_latency,
                                  unsigned int dram_latency) : hierarchy(L1_config, L2_config)
 {
-    latencies[ExclusiveCache::ACCESS_STATUS::L1_HIT] = L1_latency;
-    latencies[ExclusiveCache::ACCESS_STATUS::L1_MISS_L2_HIT] = L2_latency;
-    latencies[ExclusiveCache::ACCESS_STATUS::L1_MISS_L2_MISS] = dram_latency;
+    latencies[static_cast<int>(ExclusiveCache::ACCESS_STATUS::L1_HIT)] = L1_latency;
+    latencies[static_cast<int>(ExclusiveCache::ACCESS_STATUS::L1_MISS_L2_HIT)] = L2_latency;
+    latencies[static_cast<int>(ExclusiveCache::ACCESS_STATUS::L1_MISS_L2_MISS)] = dram_latency;
 }
 
 unsigned CacheController::access(addr_t address, unsigned tid)
 {
     const ExclusiveCache::ACCESS_STATUS status = hierarchy.access(address, tid);
 
-    return latencies[status];
+    return latencies[static_cast<int>(status)];
 }
